Change ArrayList to List, moved general function to FunktionalTestBase
